requirements.txt
requests
selenium
colorama
beautifulsoup4
python-dotenv